pyproject.toml
[project]
name = "talk2bi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "databricks-sql-connector>=3.0.0",
    "databricks-sqlalchemy>=1.0.0",
    "ddgs>=9.10.0",
    "dotenv>=0.9.9",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "streamlit>=1.54.0",
    "watchdog>=6.0.0",
]
